backend/requirements.txt
Flask==3.0.3
PyMySQL==1.1.1
cryptography==42.0.5
pyjwt==2.10.1
pytest==8.3.4
